Fixed non-ascii characters, this file was probably checked in with the wrong encoding. Fixes bug #670739. (And what a terrible idea to use non-ascii chars for macro expansion to begin with :-( )
TODO: for Python 2.3, the source encoding can and should be defined in the
source. There is a new directive for that, but I don't know the details
off-hand. Without it, Python 2.3 warns that it uses non ascii characters
without being able to establish the encoding (which I think then defaults
to latin-1, to be compatible with Python 2.2 and earlier). This should
only affect the template documents, in all other sources we should avoid
non-ascii characters.


git-svn-id: http://svn.red-bean.com/pyobjc/trunk/pyobjc@403 f55f28a5-9edb-0310-a011-a803cfcd5d25